Refactored policies out into another util
