+++ modules/core/client/controllers/participant-portal.client.controller.js
@@ -27,7 +27,6 @@ angular.module('core').controller('ParticipantPortalController', ['$scope','$htt
           'x-access-token': $scope.authToken
         }
       };
-      console.log($scope.header);
 
       $scope.sessions.getUserSessions($scope.user._id)
         .then((results) => {
