+++ modules/utils/server/authUtils.js
@@ -1,16 +1,17 @@
 const jwt = require('jsonwebtoken');
 const lodash = require('lodash');
+const policies = require('./userPolicyUtils');
 
 exports.authUser = function(req, res, next) {
 
   const reqPath = req.path;
   console.log('tw req', reqPath);
 
-  if (lodash.contains(hostRoutes, reqPath)) {
+  if (lodash.contains(policies.hostRoutes, reqPath)) {
     next();
   }
 
-  const isSecured = allSecureRoutes.some((route) => {
+  const isSecured = policies.secureCommonRoutes.some((route) => {
     if (reqPath.includes(route)) return true;
   });
 
@@ -65,11 +66,11 @@ const parseAuthToken = function(req) {
 const checkRolePermissions = (role, reqMethod, reqPath) => {
   switch (role) {
     case 'participant':
-      return checkPermissions(participant, reqMethod, reqPath);
+      return checkPermissions(policies.roles.participant, reqMethod, reqPath);
     case 'researcher':
-      return checkPermissions(researcher, reqMethod, reqPath);
+      return checkPermissions(policies.roles.researcher, reqMethod, reqPath);
     case 'faculty':
-      return checkPermissions(faculty, reqMethod, reqPath);
+      return checkPermissions(policies.roles.faculty, reqMethod, reqPath);
     case 'admin':
       return true;
     default:
@@ -80,11 +81,9 @@ const checkRolePermissions = (role, reqMethod, reqPath) => {
 const checkPermissions = (rolePermissions, reqMethod, reqPath) => {
   let isAllowed = false;
 
-  return rolePermissions.roles.some((role) => {
-    console.log('check perm role', role);
-
-    isAllowed = role[reqMethod].some((pathRegex) => {
-      console.log('check path in role', pathRegex, reqPath, pathRegex.test(reqPath));
+  return rolePermissions.permissions.some((permission) => {
+    isAllowed = permission[reqMethod].some((pathRegex) => {
+      console.log('check path in permission', pathRegex, reqPath, pathRegex.test(reqPath));
       if (pathRegex.test(reqPath)) return true;
     });
 
@@ -101,124 +100,3 @@ const expiredTokenErr = {
   code: 401,
   message: 'Your session has expired, please log back in.'
 };
-
-// Routes that are used by host and require special access
-const hostRoutes = [
-  '/api/sessions/reminderEmails'
-];
-
-const secureBasicRoutes = [
-  '/settings',
-  '/profile',
-  '/accounts',
-  '/picture',
-  '/api/studies',
-  '/api/studySessions',
-  '/api/courses',
-  '/api/sessions'
-];
-
-const participantRole = {
-  GET: [
-    /^\/api\/sessions\/user\/\w*$/,
-    /^\/api\/studySessions\/signup\/\w*\/\w*$/,
-    /^\/api\/courses\/$/,
-    /^\/api\/studies\/$/,
-  ],
-  PUT: [
-
-  ],
-  POST: [
-    /^\/api\/studySession\/signup$/
-  ],
-  DELETE: [
-    /^\/api\/sessions\/\w*$/
-  ]
-};
-
-
-const researcherRole = {
-  GET: [
-    /^\/api\/studies\/\w*$/,
-    /^\/api\/sessions\/user\/\w*$/,
-    /^\/api\/studies\/user\/\w*$/,
-    /^\/api\/studySessions\/\w*$/,
-  ],
-  PUT: [
-    /^\/api\/studies\/\w*$/,
-    /^\/api\/sessions\/attend\/\w*$/,
-    /^\/api\/sessions\/compensate\/\w*$/,
-    /^\/api\/studies\/close\/\w*$/,
-    /^\/api\/studies\/reopen\/\w*$/,
-    /^\/api\/sessions\/approveUser\/\w*$/,
-    /^\/api\/sessions\/denyUser\/\w*$/
-  ],
-  POST: [
-    /^\/api\/studies\/$/,
-    /^\/api\/sessions\/create\/\w*$/
-  ],
-  DELETE: [
-    /^\/api\/sessions\/\w*$/,
-    /^\/api\/sessions\/cancel\/\w*$/
-  ]
-};
-
-
-const facultyRole = {
-  GET: [
-    /^\/api\/sessions\/course\/\w*$/,
-    /^\/api\/courses\/$/,
-  ],
-  PUT: [
-  ],
-  POST: [
-    /^\/api\/courses\/\w*$/
-  ],
-  DELETE: [
-  ]
-};
-const adminRole = {
-  GET: [
-    '/api/sessions/user/',
-    '/api/',
-    '/api/studySessions/signup/',
-    '/api/courses'
-  ],
-  PUT: [
-
-  ],
-  POST: [
-    '/api/studySession/signup'
-  ],
-  DELETE: [
-    '/api/sessions/'
-  ]
-};
-
-const participant = {
-  roles: [participantRole]
-};
-
-const researcher = {
-  roles: [researcherRole, participantRole]
-};
-
-const faculty = {
-  roles: [facultyRole, researcherRole, participantRole]
-};
-
-exports.generateCancellationToken = function(object) {
-  const token = jwt.sign(object, process.env.JWT);
-  return token;
-};
-
-exports.parseCancellationToken = function(token) {
-  const object = jwt.verify(token, process.env.JWT);
-  return object;
-};
-
-const allSecureRoutes = secureBasicRoutes;
-
-exports.secureBasicRoutes = secureBasicRoutes;
-
-exports.allSecureRoutes = allSecureRoutes;
